requirements.txt
# Adaptive AI Career Architect - Requirements
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI/ML - Hugging Face
transformers>=4.35.0
huggingface-hub>=0.19.0
torch>=2.0.0
sentence-transformers>=2.2.0

# LangChain for Agentic Approach
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0

# Web Scraping & APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0

# GitHub Integration
PyGithub>=2.1.0

# Database
sqlalchemy>=2.0.0
sqlite3-api>=2.0.0

# Text Processing
nltk>=3.8.0
spacy>=3.7.0
textstat>=0.7.3

# React Portfolio Generation
jinja2>=3.1.0

# Metrics & Analytics
plotly>=5.18.0
altair>=5.2.0

# Utils
tqdm>=4.66.0
loguru>=0.7.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
